Some more file cleanup
